+++ src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { Layout } from "./Layout";
+export { Page } from "./Page";
