+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export { ErrorPage } from "./ErrorPage";
+export { HomePage } from "./HomePage";
+export { HotelPage } from "./HotelPage";
